# opt: let memdb cache last traversed node