--- a/internal/locate/region_cache_test.go
+++ b/internal/locate/region_cache_test.go
@@ -2689,3 +2689,147 @@ func (s *testRegionCacheSuite) testBatchScanRegions() {
 		s.cluster.Split(newID1, newID2, []byte("e1"), newPeers, newPeers[0])
 	}, []uint64{regions[1], regions[3], regions[4], newID1, regions[5], regions[6]})
 }
+
+func (s *testRegionCacheSuite) TestRangesAreCoveredCheck() {
+	check := func(ranges []string, regions []string, limit int, expect bool) {
+		rs := make([]pd.KeyRange, 0, len(ranges)/2)
+		for i := 0; i < len(ranges); i += 2 {
+			rs = append(rs, pd.KeyRange{StartKey: []byte(ranges[i]), EndKey: []byte(ranges[i+1])})
+		}
+		rgs := make([]*pd.Region, 0, len(regions))
+		for i := 0; i < len(regions); i += 2 {
+			rgs = append(rgs, &pd.Region{Meta: &metapb.Region{
+				StartKey: []byte(regions[i]),
+				EndKey:   []byte(regions[i+1]),
+			}})
+		}
+		s.Equal(expect, regionsHaveGapInRanges(rs, rgs, limit))
+	}
+	boundCases := [][]string{
+		{"a", "c"},
+		{"a", "b", "b", "c"},
+		{"a", "a1", "a1", "b", "b", "b1", "b1", "c"},
+	}
+	for _, boundCase := range boundCases {
+		// positive
+		check(boundCase, []string{"a", "c"}, -1, false)
+		check(boundCase, []string{"a", ""}, -1, false)
+		check(boundCase, []string{"", "c"}, -1, false)
+		// negative
+		check(boundCase, []string{"a", "b"}, -1, true)
+		check(boundCase, []string{"b", "c"}, -1, true)
+		check(boundCase, []string{"b", ""}, -1, true)
+		check(boundCase, []string{"", "b"}, -1, true)
+		// positive
+		check(boundCase, []string{"a", "b", "b", "c"}, -1, false)
+		check(boundCase, []string{"", "b", "b", "c"}, -1, false)
+		check(boundCase, []string{"a", "b", "b", ""}, -1, false)
+		check(boundCase, []string{"", "b", "b", ""}, -1, false)
+		// negative
+		check(boundCase, []string{"a", "b", "b1", "c"}, -1, true)
+		check(boundCase, []string{"", "b", "b1", "c"}, -1, true)
+		check(boundCase, []string{"a", "b", "b1", ""}, -1, true)
+		check(boundCase, []string{"", "b", "b1", ""}, -1, true)
+		check(boundCase, []string{}, -1, true)
+	}
+
+	nonContinuousCases := [][]string{
+		{"a", "b", "c", "d"},
+		{"a", "b1", "b1", "b", "c", "d"},
+		{"a", "b", "c", "c1", "c1", "d"},
+		{"a", "b1", "b1", "b", "c", "c1", "c1", "d"},
+	}
+	for _, nonContinuousCase := range nonContinuousCases {
+		// positive
+		check(nonContinuousCase, []string{"a", "d"}, -1, false)
+		check(nonContinuousCase, []string{"", "d"}, -1, false)
+		check(nonContinuousCase, []string{"a", ""}, -1, false)
+		check(nonContinuousCase, []string{"", ""}, -1, false)
+		// negative
+		check(nonContinuousCase, []string{"a", "b"}, -1, true)
+		check(nonContinuousCase, []string{"b", "c"}, -1, true)
+		check(nonContinuousCase, []string{"c", "d"}, -1, true)
+		check(nonContinuousCase, []string{"", "b"}, -1, true)
+		check(nonContinuousCase, []string{"c", ""}, -1, true)
+	}
+
+	unboundCases := [][]string{
+		{"", ""},
+		{"", "b", "b", ""},
+		{"", "a1", "a1", "b", "b", "b1", "b1", ""},
+	}
+	for _, unboundCase := range unboundCases {
+		// positive
+		check(unboundCase, []string{"", ""}, -1, false)
+		// negative
+		check(unboundCase, []string{"a", "c"}, -1, true)
+		check(unboundCase, []string{"a", ""}, -1, true)
+		check(unboundCase, []string{"", "c"}, -1, true)
+		// positive
+		check(unboundCase, []string{"", "b", "b", ""}, -1, false)
+		// negative
+		check(unboundCase, []string{"", "b", "b1", ""}, -1, true)
+		check(unboundCase, []string{"a", "b", "b", ""}, -1, true)
+		check(unboundCase, []string{"", "b", "b", "c"}, -1, true)
+		check(unboundCase, []string{}, -1, true)
+	}
+
+	// test half bounded ranges
+	check([]string{"", "b"}, []string{"", "a"}, -1, true)
+	check([]string{"", "b"}, []string{"", "a"}, 1, false) // it's just limitation reached
+	check([]string{"", "b"}, []string{"", "a"}, 2, true)
+	check([]string{"a", ""}, []string{"b", ""}, -1, true)
+	check([]string{"a", ""}, []string{"b", ""}, 1, true)
+	check([]string{"a", ""}, []string{"b", "c"}, 1, true)
+	check([]string{"a", ""}, []string{"a", ""}, -1, false)
+}
+
+func (s *testRegionCacheSuite) TestScanRegionsWithGaps() {
+	// Split at "a", "c", "e"
+	// nil --- 'a' --- 'c' --- 'e' --- nil
+	// <-  0  -> <- 1 -> <- 2 -> <- 3 -->
+	regions := s.cluster.AllocIDs(3)
+	regions = append([]uint64{s.region1}, regions...)
+
+	peers := [][]uint64{{s.peer1, s.peer2}}
+	for i := 0; i < 3; i++ {
+		peers = append(peers, s.cluster.AllocIDs(2))
+	}
+
+	for i := 0; i < 3; i++ {
+		s.cluster.Split(regions[i], regions[i+1], []byte{'a' + 2*byte(i)}, peers[i+1], peers[i+1][0])
+	}
+
+	// the last region is not reported to PD yet
+	getRegionIDsWithInject := func(fn func() ([]*Region, error)) []uint64 {
+		s.cache.clear()
+		err := failpoint.Enable("tikvclient/mockSplitRegionNotReportToPD", fmt.Sprintf(`return(%d)`, regions[2]))
+		s.Nil(err)
+		resCh := make(chan []*Region)
+		errCh := make(chan error)
+		go func() {
+			rs, err := fn()
+			errCh <- err
+			resCh <- rs
+		}()
+		time.Sleep(time.Second)
+		failpoint.Disable("tikvclient/mockSplitRegionNotReportToPD")
+		s.Nil(<-errCh)
+		rs := <-resCh
+		regionIDs := make([]uint64, 0, len(rs))
+		for _, r := range rs {
+			regionIDs = append(regionIDs, r.GetID())
+		}
+		return regionIDs
+	}
+
+	scanRegionRes := getRegionIDsWithInject(func() ([]*Region, error) {
+		return s.cache.BatchLoadRegionsWithKeyRange(s.bo, []byte(""), []byte(""), 10)
+	})
+	s.Equal(scanRegionRes, regions)
+
+	batchScanRegionRes := getRegionIDsWithInject(func() ([]*Region, error) {
+		return s.cache.BatchLoadRegionsWithKeyRanges(s.bo, []pd.KeyRange{{StartKey: []byte{}, EndKey: []byte{}}}, 10)
+	})
+	s.Equal(batchScanRegionRes, regions)
+}