This PR:
```
go test -run=^$ -bench=Record -benchtime=10s
goos: darwin
goarch: arm64
pkg: github.com/pingcap/tidb/pkg/table/tables
BenchmarkAddRecordInPipelinedDML-10       	    1381	   8611614 ns/op	       861.2 ns/record
BenchmarkRemoveRecordInPipelinedDML-10    	    3180	   3800922 ns/op	       380.1 ns/record
BenchmarkUpdateRecordInPipelinedDML-10    	    2402	   5006901 ns/op	       500.7 ns/record
PASS
ok  	github.com/pingcap/tidb/pkg/table/tables	108.504s
```
Master:
```
go test -run=^$ -bench=Record -benchtime=10s
goos: darwin
goarch: arm64
pkg: github.com/pingcap/tidb/pkg/table/tables
BenchmarkAddRecordInPipelinedDML-10       	     962	  12420272 ns/op	      1242 ns/record
BenchmarkRemoveRecordInPipelinedDML-10    	    2365	   4975164 ns/op	       497.5 ns/record
BenchmarkUpdateRecordInPipelinedDML-10    	    1908	   6252301 ns/op	       625.2 ns/record
PASS
ok  	github.com/pingcap/tidb/pkg/table/tables	111.274s
```